refactor(domain): clean up game module tests
